requirements.txt
streamlit
scikit-learn
pandas
matplotlib
opencv-python-headless
pillow
